+++ lib/intercepted_http.dart
@@ -3,4 +3,5 @@ library;
 
 export 'src/http_exception.dart';
 export 'src/http_interceptor.dart';
+export 'src/http_status_code.dart';
 export 'src/intercepted_http.dart';